Move conjested initial state to separate file
