+++ src/reducers/stateData/bookIntialStateData.js
@@ -1,0 +1,37 @@
+const bookA = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Divergent',
+  category: 'Action',
+};
+
+const bookB = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Maze Runner',
+  category: 'Action',
+};
+
+const bookC = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Enders Game',
+  category: 'Sci-Fi',
+};
+
+const bookD = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Harry Potter',
+  category: 'Sci-Fi',
+};
+
+const bookE = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Bicos House',
+  category: 'Horror',
+};
+
+const bookF = {
+  id: Math.floor(Math.random() * 100) + 1,
+  title: 'Timeless',
+  category: 'Biography',
+};
+
+export const initialState = [bookA, bookB, bookC,bookD,bookE,bookF];
